requirements.txt
sentence-transformers==2.7.0
torch>=2.1.0,<2.3
numpy>=1.24,<2.0
pyyaml>=6.0
scikit-learn>=1.3
